fixed remarks
